+++ src/main/java/com/chandara/phoneshop/Service/BrandService.java
@@ -8,9 +8,9 @@ import com.chandara.phoneshop.entity.Brand;
 
 public interface BrandService {
 	Brand create(Brand brand);
-	Brand getById(Integer id);
-	Brand update(Integer id,Brand brandUpdate);
-	void delete(Integer id);
+	Brand getById(Long id);
+	Brand update(Long id,Brand brandUpdate);
+	void delete(Long id);
 	List<Brand> getBrands();
 	List<Brand> getBrands(String name);
 	//List<Brand> getBrands(Map<String, String> params);
