+++ src/main/java/com/chandara/phoneshop/repository/ModelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.chandara.phoneshop.entity.Model;
 
 
-public interface ModelRepository extends JpaRepository<Model, Integer> {
-	List<Model> findByBrandId(Integer brandId); 
+public interface ModelRepository extends JpaRepository<Model, Long> {
+	List<Model> findByBrandId(Long brandId);
 	}
